+++ src/components/layout/Layout.tsx
@@ -1,8 +1,9 @@
-import { Header } from "../header";
-import { Navbar } from "../navbar";
 import { PropsWithChildren } from "react";
 
 import * as S from "./styled";
+
+import { Header } from "./header";
+import { Navbar } from "./navbar";
 
 export default function Layout({ children }: PropsWithChildren) {
   return (
