+++ src/components/layout/navbar/Navbar.tsx
@@ -1,10 +1,11 @@
+import { useRouter } from "next/router";
+
 import { useScrollDirection } from "@/hooks";
 
 import { NAV_MENUS } from "@/assets/static";
 
 import * as S from "./Navbar.styled";
 import NavItem from "./item/NavItem";
-import { useRouter } from "next/router";
 
 export default function Navbar() {
   const { pathname } = useRouter();
